requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-frontmatter==1.1.0 